feat: made isDiscussionTimeChecker functions compatible with new database architecture of upcomingRoom's database
